+++ .github/dependabot.yml
@@ -1,4 +1,4 @@
-# To get started with Dependabot version updates, you'll need to specify which
+﻿# To get started with Dependabot version updates, you'll need to specify which
 # package ecosystems to update and where the package manifests are located.
 # Please see the documentation for all configuration options:
 # https://help.github.com/github/administering-a-repository/configuration-options-for-dependency-updates
@@ -12,7 +12,7 @@ updates:
     open-pull-requests-limit: 5
     reviewers:
       - EmotionChild
-    target-branch: "main"
+    target-branch: "development"
     labels:
       - "nuget"
       - "dependencies"
